pulseai/member2_voice_agent/requirements.txt
Flask==3.0.0
Werkzeug==3.0.0
openai==1.12.0
google-cloud-speech==2.21.0
pydantic==2.5.2
spacy==3.7.2
python-dotenv==1.0.0
pytest==7.4.3
requests==2.31.0
